move files around
create sub-folders for components:
- app: singleton components. just the app shell for now
- three: 3d components
- common: random stuff
