--- frontend/components/World.module.scss
@@ -1,10 +1,0 @@
-.wrapper {
-  z-index: -1;
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-}
