+++ frontend/components/app/Shell.module.scss
@@ -15,3 +15,14 @@
   display: block;
   position: relative;
 }
+
+.world {
+  z-index: -1;
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+}
