+++ frontend/components/three/Earth.tsx
